test: build the app without starting it, and sweep the privacy guards (closes #46)
The promise this project makes to strangers on the internet — private is
private, server-side, with no exception for the administrator — had no
test coverage at all, and could not have any. index.ts built the app and
called listen() at module scope, so importing it started a server.

So the HTTP surface moves to buildApp() in app.ts: hooks, plugins,
routes, and nothing that touches the world. index.ts keeps what starting
a server means — migrations, the pruner, demo init, the port, the mail
poller, signal handlers. Nothing about the surface itself changed, which
the boot check confirms: health, setup, a real write, the CSP and
rate-limit headers, and redactUrl still masking ?token=… in the log.

On top of that, a sweep: Alice creates something private, Bob asks for it
by id, and the answer must be 404 — not 200, and not 403, which would
confirm the thing exists. Notes, personal accounts, personal calendars,
across read, list, edit, delete, search and version history, plus the
unauthenticated case.

It earned itself on the first run. Two delete routes turned out to take
an id and nothing else:

- accounts, already filed as #46;
- calendars, which nobody had noticed.

Both now do the same lookup their neighbouring archive and patch handlers
already did. An audit of the remaining fourteen delete routes found no
third: the ones without an owner check are shared by design (categories,
projects, folders, tasks) or admin-gated (invites).

Removing either guard turns the sweep red, which is the point of writing
it this way round.
